third_party/blink/web_tests/external/wpt/geolocation-sensor/idlharness.https.window-expected.txt

This is a testharness.js-based test.
PASS idl_test setup
PASS idl_test validation
FAIL GeolocationSensor interface: existence and properties of interface object assert_own_property: self does not have own property "GeolocationSensor" expected property "GeolocationSensor" missing
FAIL GeolocationSensor interface object length assert_own_property: self does not have own property "GeolocationSensor" expected property "GeolocationSensor" missing
FAIL GeolocationSensor interface object name assert_own_property: self does not have own property "GeolocationSensor" expected property "GeolocationSensor" missing
FAIL GeolocationSensor interface: existence and properties of interface prototype object assert_own_property: self does not have own property "GeolocationSensor" expected property "GeolocationSensor" missing
FAIL GeolocationSensor interface: existence and properties of interface prototype object's "constructor" property assert_own_property: self does not have own property "GeolocationSensor" expected property "GeolocationSensor" missing
FAIL GeolocationSensor interface: existence and properties of interface prototype object's @@unscopables property assert_own_property: self does not have own property "GeolocationSensor" expected property "GeolocationSensor" missing
FAIL GeolocationSensor interface: operation read(ReadOptions) assert_own_property: self does not have own property "GeolocationSensor" expected property "GeolocationSensor" missing
FAIL GeolocationSensor interface: attribute latitude assert_own_property: self does not have own property "GeolocationSensor" expected property "GeolocationSensor" missing
FAIL GeolocationSensor interface: attribute longitude assert_own_property: self does not have own property "GeolocationSensor" expected property "GeolocationSensor" missing
FAIL GeolocationSensor interface: attribute altitude assert_own_property: self does not have own property "GeolocationSensor" expected property "GeolocationSensor" missing
FAIL GeolocationSensor interface: attribute accuracy assert_own_property: self does not have own property "GeolocationSensor" expected property "GeolocationSensor" missing
FAIL GeolocationSensor interface: attribute altitudeAccuracy assert_own_property: self does not have own property "GeolocationSensor" expected property "GeolocationSensor" missing
FAIL GeolocationSensor interface: attribute heading assert_own_property: self does not have own property "GeolocationSensor" expected property "GeolocationSensor" missing
FAIL GeolocationSensor interface: attribute speed assert_own_property: self does not have own property "GeolocationSensor" expected property "GeolocationSensor" missing
FAIL GeolocationSensor must be primary interface of new GeolocationSensor assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: GeolocationSensor is not defined"
FAIL Stringification of new GeolocationSensor assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: GeolocationSensor is not defined"
FAIL GeolocationSensor interface: new GeolocationSensor must inherit property "read(ReadOptions)" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: GeolocationSensor is not defined"
FAIL GeolocationSensor interface: calling read(ReadOptions) on new GeolocationSensor with too few arguments must throw TypeError assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: GeolocationSensor is not defined"
FAIL GeolocationSensor interface: new GeolocationSensor must inherit property "latitude" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: GeolocationSensor is not defined"
FAIL GeolocationSensor interface: new GeolocationSensor must inherit property "longitude" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: GeolocationSensor is not defined"
FAIL GeolocationSensor interface: new GeolocationSensor must inherit property "altitude" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: GeolocationSensor is not defined"
FAIL GeolocationSensor interface: new GeolocationSensor must inherit property "accuracy" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: GeolocationSensor is not defined"
FAIL GeolocationSensor interface: new GeolocationSensor must inherit property "altitudeAccuracy" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: GeolocationSensor is not defined"
FAIL GeolocationSensor interface: new GeolocationSensor must inherit property "heading" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: GeolocationSensor is not defined"
FAIL GeolocationSensor interface: new GeolocationSensor must inherit property "speed" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: GeolocationSensor is not defined"
Harness: the test ran to completion.
